Introduce useInitSearch
